Changes to the dynamic_brightcove.js, brightcove_admin.php, and brightcove.php files.
brightcove_admin.php:

-simple typo fix

brightcove.php:

-some simple clean up

dynamic-brightcove.js:

-moved the default helper functions to the top of the file so they
could be used in the replaceTokens function

-added a 3rd parameter to the replaceTokens function so that the type
of player could be distinguished

-fixed the way the shortcode variable was being made to still work even
if the height or width were "undefined"

-took out the base url plus "services/library/&" since it seems the api
auto concatenates that part of the url on to the front of the url that
is sent off to the api

-added some spacing to make the code cleaner and more representable

-changed the conditional to check and see if any items were brought
back from the API instead of checking to see how many pages there were

-added some more things inside the "totalCount == 0" conditional to
make the plugin function better

-added the removal of the video/playlist placeholder to the
generateHTMLForPlaylist function so it will be removed when the "See
All Videos/See All Playlists" button is clicked

-Added the createShowAllVideosButton function right after the
displayPagedVideoSearchResults function so that the
"See All Videos" button will show when a search was made for videos

-made the "search videos" value blank when the "See All Videos" button
is  clicked

-added a conditional to check and see if either the width or the height
came through as undefined. If they did, then the default width or
default height will be used.